+++ backend/src/test/java/com/woowacourse/f12/application/inventoryproduct/InventoryProductServiceTest.java
@@ -9,30 +9,20 @@ import static com.woowacourse.f12.support.ProductFixture.SOFTWARE_1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.refEq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.verify;
-import static org.mockito.Mockito.times;
 
 import com.woowacourse.f12.domain.inventoryproduct.InventoryProduct;
 import com.woowacourse.f12.domain.inventoryproduct.InventoryProductRepository;
 import com.woowacourse.f12.domain.member.Member;
 import com.woowacourse.f12.domain.member.MemberRepository;
-import com.woowacourse.f12.domain.product.Product;
-import com.woowacourse.f12.domain.review.Review;
 import com.woowacourse.f12.domain.review.ReviewRepository;
 import com.woowacourse.f12.dto.request.inventoryproduct.ProfileProductRequest;
 import com.woowacourse.f12.dto.response.inventoryproduct.InventoryProductResponse;
 import com.woowacourse.f12.dto.response.inventoryproduct.InventoryProductsResponse;
-import com.woowacourse.f12.dto.response.review.ReviewWithProductResponse;
 import com.woowacourse.f12.exception.badrequest.DuplicatedProfileProductCategoryException;
 import com.woowacourse.f12.exception.badrequest.InvalidProfileProductCategoryException;
 import com.woowacourse.f12.exception.badrequest.NotUpdatableException;
-import com.woowacourse.f12.exception.notfound.InventoryProductNotFoundException;
-import com.woowacourse.f12.exception.notfound.ReviewNotFoundException;
-import com.woowacourse.f12.support.ReviewFixtures;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
@@ -174,60 +164,6 @@ class InventoryProductServiceTest {
                         .containsOnly(InventoryProductResponse.from(inventoryProduct)),
                 () -> verify(memberRepository).existsById(memberId),
                 () -> verify(inventoryProductRepository).findByMemberId(memberId)
-        );
-    }
-
-    @Test
-    void 인벤토리_아이디로_리뷰를_조회한다() {
-        // given
-        Review review = ReviewFixtures.REVIEW_RATING_1.작성(1L, KEYBOARD_1.생성(1L), CORINNE.생성(1L));
-        given(inventoryProductRepository.findById(any(Long.class)))
-                .willReturn(Optional.of(SELECTED_INVENTORY_PRODUCT.생성(CORINNE.생성(1L), KEYBOARD_1.생성(1L))));
-        given(reviewRepository.findByMemberAndProduct(any(Member.class), any(Product.class)))
-                .willReturn(Optional.of(review));
-
-        // when
-        ReviewWithProductResponse reviewResponse = inventoryProductService.findReviewById(1L);
-
-        // then
-        assertAll(
-                () -> assertThat(reviewResponse).usingRecursiveComparison()
-                        .isEqualTo(ReviewWithProductResponse.from(review)),
-                () -> verify(inventoryProductRepository).findById(eq(1L)),
-                () -> verify(reviewRepository).findByMemberAndProduct(refEq(CORINNE.생성(1L)), refEq(KEYBOARD_1.생성(1L)))
-        );
-    }
-
-    @Test
-    void 인벤토리_아이디로_리뷰를_조회할때_등록된_장비가_존재하지_않는다면_예외가_발생한다() {
-        // given
-        given(inventoryProductRepository.findById(any(Long.class)))
-                .willReturn(Optional.empty());
-
-        // when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> inventoryProductService.findReviewById(1L))
-                        .isInstanceOf(InventoryProductNotFoundException.class),
-                () -> verify(inventoryProductRepository).findById(eq(1L)),
-                () -> verify(reviewRepository, times(0)).findByMemberAndProduct(any(Member.class), any(Product.class))
-        );
-    }
-
-    @Test
-    void 인벤토리_아이디로_리뷰를_조회할때_리뷰가_존재하지_않는다면_예외가_발생한다() {
-        //given
-        Review review = ReviewFixtures.REVIEW_RATING_1.작성(1L, KEYBOARD_1.생성(1L), CORINNE.생성(1L));
-        given(inventoryProductRepository.findById(any(Long.class)))
-                .willReturn(Optional.of(SELECTED_INVENTORY_PRODUCT.생성(CORINNE.생성(1L), KEYBOARD_1.생성(1L))));
-        given(reviewRepository.findByMemberAndProduct(any(Member.class), any(Product.class)))
-                .willReturn(Optional.empty());
-
-        // when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> inventoryProductService.findReviewById(1L)).isInstanceOf(
-                        ReviewNotFoundException.class),
-                () -> verify(inventoryProductRepository).findById(eq(1L)),
-                () -> verify(reviewRepository).findByMemberAndProduct(refEq(CORINNE.생성(1L)), refEq(KEYBOARD_1.생성(1L)))
         );
     }
 }
