+++ backend/src/test/java/com/woowacourse/f12/presentation/inventoryproduct/InventoryProductControllerTest.java
@@ -3,11 +3,9 @@ package com.woowacourse.f12.presentation.inventoryproduct;
 import static com.woowacourse.f12.support.InventoryProductFixtures.SELECTED_INVENTORY_PRODUCT;
 import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
 import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_1;
-import static com.woowacourse.f12.support.ReviewFixtures.REVIEW_RATING_1;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willDoNothing;
 import static org.mockito.BDDMockito.willThrow;
@@ -25,7 +23,6 @@ import com.woowacourse.f12.domain.inventoryproduct.InventoryProduct;
 import com.woowacourse.f12.domain.member.Member;
 import com.woowacourse.f12.dto.request.inventoryproduct.ProfileProductRequest;
 import com.woowacourse.f12.dto.response.inventoryproduct.InventoryProductsResponse;
-import com.woowacourse.f12.dto.response.review.ReviewWithProductResponse;
 import com.woowacourse.f12.exception.badrequest.DuplicatedProfileProductCategoryException;
 import com.woowacourse.f12.exception.badrequest.InvalidProfileProductCategoryException;
 import com.woowacourse.f12.exception.notfound.InventoryProductNotFoundException;
@@ -249,23 +246,5 @@ class InventoryProductControllerTest {
         assertAll(
                 () -> verify(inventoryProductService).findByMemberId(memberId)
         );
-    }
-
-    @Test
-    void 인벤토리_아이디로_리뷰_조회_성공() throws Exception {
-        // given
-        Long inventoryId = 1L;
-        given(inventoryProductService.findReviewById(inventoryId))
-                .willReturn(ReviewWithProductResponse.from(REVIEW_RATING_1.작성(KEYBOARD_1.생성(), CORINNE.생성())));
-
-        // when
-        mockMvc.perform(
-                        get("/api/v1/inventoryProducts/" + inventoryId + "/reviews")
-                )
-                .andExpect(status().isOk())
-                .andDo(print());
-
-        // then
-        verify(inventoryProductService).findReviewById(eq(inventoryId));
     }
 }
